Time: 0 ms (100.00%) | Memory: 8.4 MB (91.08%) - LeetSync

+++ 3379-score-of-a-string/score-of-a-string.cpp
@@ -3,7 +3,7 @@ public:
     int scoreOfString(string s) {
         int sum = 0;
         for (int i = 1; i < s.length(); i++)
-            sum += abs((int)(s[i-1]) - (int)(s[i]));
+            sum += abs(s[i-1] - s[i]);
 
         return sum;
     }
